requirements.txt
h5py==2.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
mtcnn==0.1.0
numpy==1.16.0
opencv-python==4.2.0.32
pickleshare==0.7.5
Pillow==7.0.0
protobuf==3.11.3
requests==2.23.0
scikit-learn==0.20.2
scipy==1.4.1
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
zipp==3.1.0
